backend/requirements.txt
anyio==3.4.0
appdirs==1.4.4
asgiref==3.4.1
CacheControl==0.12.10
cached-property==1.5.2
cachy==0.3.0
certifi==2021.10.8
charset-normalizer==2.0.9
cleo==1.0.0a4
click==8.0.3
crashtest==0.3.1
distlib==0.3.4
entrypoints==0.3
fastapi==0.70.1
filelock==3.4.2
graphql-core==3.1.7
greenlet==1.1.2
h11==0.12.0
html5lib==1.1
httptools==0.3.0
idna==3.3
importlib-metadata==4.10.0
keyring==23.4.0
lockfile==0.12.2
msgpack==1.0.3
packaging==20.9
pexpect==4.8.0
pkginfo==1.8.2
poetry==1.2.0a2
poetry-core==1.1.0a6
ptyprocess==0.7.0
pydantic==1.8.2
Pygments==2.10.0
pylev==1.4.0
pyparsing==3.0.6
python-dateutil==2.8.2
python-dotenv==0.19.2
python-multipart==0.0.5
PyYAML==6.0
requests==2.26.0
requests-toolbelt==0.9.1
sentinel==0.3.0
shellingham==1.4.0
six==1.16.0
sniffio==1.2.0
starlette==0.16.0
strawberry-graphql @ https://github.com/deknowny/strawberry/archive/v0.94.0a1.zip
tomlkit==0.8.0
typing_extensions==4.0.1
urllib3==1.26.7
uvicorn==0.16.0
virtualenv==20.4.4
watchgod==0.7
webencodings==0.5.1
websockets==10.1
zipp==3.6.0
